Fixed Color ostream overload and UIObject SDL fill

+++ NerdFramework++/Color3.cpp
@@ -83,5 +83,5 @@ void Color3::writeToPixel(void* pixel) const {
 }
 
 std::ostream& operator<<(std::ostream& stream, const Color3& rhs) {
-	return stream << "{ r: " << rhs.r << ", g: " << rhs.g << ", b: " << rhs.b << " }";
+	return stream << "{ r: " << (short)rhs.r << ", g: " << (short)rhs.g << ", b: " << (short)rhs.b << " }";
 }
